K-function cross-type in 3D

Calculates an estimate of the cross-type K-function for a multitype point pattern.

K.cross.3D(X, Y, Z, X2, Y2, Z2, psz = 25, width = 1, intensity = NULL,
  intensity2 = NULL, parallel = FALSE, verbose = FALSE)

Arguments

X
X coordinate of first observed point pattern in microns.

Y
Y coordinate

Z
Z coordinate

X2
X coordinate of second observed point pattern

Y2
Y coordinate

Z2
Z coordinate

psz
pointsize used for discetization. Smaller values are more precise, but need more computation time.

width
maximum distance

intensity
intensity of first pattern. Only if $$\lambda(s)!=\lambda$$.

intensity2
intensity of second pattern

parallel
Logical. Can we use parallel computing?

verbose
Plot verbose information

Value

a list of breaks and counts.