Title Find distance to next neighbour of a specific class

Title Find distance to next neighbour of a specific class

nearestClassDistance(coord, img, class, voxelsize, step = 0)

Arguments

coord

coordinate of relevant voxel

img

image array of classes

class

class to find

voxelsize

vector of length three. size of voxel in X-/Y-/Z-direction

step

(starting) size of neighbourhood to search

Value

distance to nearest voxel of class "class"