Find all distances to next neighbour of all classes

Find all distances to next neighbour of all classes

nearestClassDistances(img, voxelsize = NULL, size = NULL, classes = 7,
  maxdist = NULL, silent = FALSE, cores = 1)

Arguments

img
Image array of classes

voxelsize
Real size of voxels in microns.

size
Real size of image in microns. Either size or voxelsize must be given.

classes
Number of classes

maxdist
Maximum distance to consider

silent
Remain silent?

cores
Number of cores available for parallel computing

Value

array with distances