Display an image stack

Display an image stack

img(x, z = NULL, ch = NULL, mask = NULL, col = "grey", low = NULL,
  up = NULL, ...)

Arguments

x
Image, 2D or 3D Matrix

z
slice to show, default: NULL, expects x to be 2d or 2d+channels

ch
channel. Default: NULL, either only one channel, rgb or channel will be assumed from col

mask
mask for image, voxel outside the mask will be transparent (default: NULL, no mask)

col
Color, either a character ("grey", "greyinvert", "red" ("r"), "green" ("g") or "blue" ("b"), rgb" for 3D matrices), a vector of character with hex rgb values or a function.

low
minimal value of shown intensity. Default: NULL: use min(x, na.rm=TRUE).

up
maximal value of shown intensity. Default: NULL: use max(x, na.rm=TRUE).

...
other parameters for graphics::image

Value

no return