Find spots based on threshold and minimum total intensity

Find spots based on threshold and minimum total intensity

spots(img, mask, thresh.offset = 0.1, min.sum.intensity = 0, zero = NA,
  max.spots = NULL, return = "full")

Arguments

img
image array

mask
mask

thresh.offset
threshold for minimum voxel intensity

min.sum.intensity
threshold for minimum total spot intensity

zero
if NA, background is set to NA, if 0,background is set to 0

max.spots
find max.spots spots with highest total intensity

return
"mask" returns binarized mask, "intensity" returns intensity for spots, zero or NA otherwise "label" return labeled (numbered) spots

Value

array